requirements.txt
torch
datasets
pillow
torchsummary
torchvision